requirement.txt
django
psycopg2
djangorestframework
django-filter
dj-rest-auth
django-cors-headers
django-rest-knox
djangorestframework_simplejwt